requirements.txt
psycopg2-binary
sqlmodel
PyYAML
pytest
fastapi
uvicorn
requests
python-dotenv